requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
bs4==0.0.1
cachelib==0.10.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chromedriver==2.24.1
click==8.1.3
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
discord.py==2.3.2
dnspython==1.15.0
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Session==0.5.0
fqdn==1.5.1
freetype-py==2.3.0
frozenlist==1.4.0
html5lib==1.1
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mecab-python3==1.0.6
mistune==3.0.1
multidict==6.0.4
mysql-connector-python==8.1.0
mysqlclient==2.1.1
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.2
notebook_shim==0.2.3
openai==0.28.0
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
pycairo==1.23.0
pycparser==2.21
pydub==0.25.1
Pygments==2.16.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qrcode==7.4.2
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
reportlab==4.0.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlPyCairo==0.2.0
rpds-py==0.10.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.5.0
unidic-lite==1.0.8
uri-template==1.3.0
urllib3==2.0.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==2.3.4
widgetsnbextension==4.0.8
yarl==1.9.2
